refactor: extract input terminal helpers
